requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11
django-cors-headers==3.2.1
django-crispy-forms==1.9.1
django-filter==2.2.0
django-formtools==2.2
django-import-export==2.1.0
django-reversion==3.0.7
djangorestframework==3.8.0
et-xmlfile==1.0.1
future==0.18.2
get==2019.4.13
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
lxml==4.5.0
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
post==2019.4.13
public==2019.4.13
PyMySQL==0.9.3
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
zipp==3.1.0
